requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0
sqlite-vec>=0.1.0
accelerate>=0.24.0
tqdm>=4.66.0